src/modular_spaceflights/pipelines/reporting/requirements.txt
matplotlib==3.5.0
pillow==8.3.2
